Crate versoview_build

VeroView Build

This is a crate to help with getting started with using verso as a webview without building it yourself

Example

To use it, first add it to your build dependency, and in your build script:

fn main() {
    versoview_build::download_and_extract_verso("output_directory").unwrap();
}

Functions

decompress_archive

Decompress the archive to the output directory, this should resulting in a versoview(.exe) in that directory

default_archive_base_url

Default archive base URL: https://github.com/versotile-org/versoview-release/releases

default_output_directory

If you don’t know where to put the versoview executable, this function gives you the target directory

download_and_extract_verso

Download and extract the pre-built versoview executable to this directory

download_archive

Download the pre-built versoview archive to the OUT_DIR and returns that path